requirements.txt
Django==3.1.4
dj-database-url==0.5.0
gunicorn==20.1.0
Unipath==1.0
python-decouple==3
Pillow==3.3.0
bleach==1.4.3
psycopg2==2.8.6
whitenoise>=3.2
django-heroku==0.3.1
dj-database-url==0.5.0
Django==3.1.4
django-allauth==0.44.0
django-cors-headers==3.7.0
django-crispy-forms @ file:
django-debug-toolbar==3.2.1
django-extensions==3.1.3
django-filter==2.4.0
django-heroku==0.3.1
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
djangorestframework==3.12.4
filelock==3.0.12
gunicorn==20.1.0
